# TicketManagementSystem/backend/TicketManagementSystem.API/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Management System API
  description: |
    API RESTful para el Sistema de Gestión de Tickets.
    
    ## Autenticación
    Esta API utiliza **JWT Bearer Token** para autenticación. Para acceder a los endpoints protegidos:
    1. Obtener un token mediante `/api/auth/login`
    2. Incluir el token en el header `Authorization: Bearer {token}`
    
    ## Rate Limiting
    - General: 200 requests/minuto
    - Login: 20 requests/minuto
    - Register: 10 requests/minuto
    
    ## Roles
    - **Admin**: Acceso completo al sistema
    - **Agent**: Gestión de tickets asignados
    - **User**: Crear y ver tickets propios
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@ticketmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
  - url: https://localhost:5001/api
    description: Servidor de desarrollo (HTTPS)
  - url: https://api.ticketmanagement.com/api
    description: Servidor de producción
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Listar usuarios
      description: Obtener lista paginada de usuarios (solo Admin)
      operationId: getUsers
      parameters:
        - name: page
          in: query
          description: Número de página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Cantidad de elementos por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: role
          in: query
          description: Filtrar por rol
          schema:
            type: string
            enum: [Admin, Agent, User]
        - name: isActive
          in: query
          description: Filtrar por estado activo
          schema:
            type: boolean
        - name: search
          in: query
          description: Buscar por nombre o email
          schema:
            type: string
      responses:
        '200':
          description: Lista de usuarios paginada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseUserDto'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Sin permisos para listar usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Obtener usuario por ID
      description: Obtener detalles de un usuario específico (solo Admin)
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Sin permisos para ver usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    put:
      tags:
        - Users
      summary: Actualizar usuario
      description: Actualizar un usuario existente (solo Admin)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
            example:
              fullName: "Juan Carlos Pérez"
              role: "Agent"
              isActive: true
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Sin permisos para modificar usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      tags:
        - Users
      summary: Eliminar usuario
      description: Eliminar un usuario (solo Admin)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: integer
      responses:
        '204':
          description: Usuario eliminado exitosamente
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Sin permisos para eliminar usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ==================== COMMENTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido mediante el endpoint `/api/auth/login`.
        
        Formato: `Bearer {token}`
        
        Ejemplo: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          description: ID único del usuario
          example: 1
        email:
          type: string
          format: email
          description: Email del usuario
          example: "user@example.com"
        fullName:
          type: string
          description: Nombre completo del usuario
          example: "Juan Pérez"
        role:
          type: string
          enum: [Admin, Agent, User]
          description: Rol del usuario
          example: "User"
        isActive:
          type: boolean
          description: Indica si la cuenta está activa
          example: true
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación
          example: "2025-11-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última actualización
          example: "2025-11-15T14:30:00Z"
...
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Nuevo email
          example: "updated@example.com"
        fullName:
          type: string
          description: Nuevo nombre completo
          example: "Juan Carlos Pérez"
        role:
          type: string
          enum: [Admin, Agent, User]
          description: Nuevo rol
          example: "Agent"
        isActive:
          type: boolean
          description: Estado de la cuenta
          example: true

    # ==================== TICKET SCHEMAS ====================
...
    PagedResponseUserDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
          description: Lista de usuarios
        totalItems:
          type: integer
          description: Total de items en la base de datos
          example: 50
        page:
          type: integer
          description: Página actual
          example: 1
        pageSize:
          type: integer
          description: Tamaño de página
          example: 10
        totalPages:
          type: integer
          description: Total de páginas
          example: 5
        hasNextPage:
          type: boolean
          description: Indica si hay página siguiente
          example: true
        hasPreviousPage:
          type: boolean
          description: Indica si hay página anterior
          example: false

    # ==================== ERROR SCHEMAS ====================
    ProblemDetails:
      type: object
      description: Formato de error según RFC 7807
      properties:
        type:
          type: string
          description: URI que identifica el tipo de error
          example: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        title:
          type: string
          description: Título del error
          example: "Bad Request"
        status:
          type: integer
          description: Código HTTP
          example: 400
        detail:
          type: string
          description: Descripción detallada del error
          example: "El campo 'email' es requerido"
        instance:
          type: string
          description: URI de la instancia del error
          example: "/api/auth/login"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Errores de validación por campo
          example:
            email:
              - "El campo 'email' es requerido"
              - "El formato del email es inválido"
            password:
              - "La contraseña debe tener al menos 6 caracteres"
        traceId:
          type: string
          description: ID de traza para debugging
          example: "00-abc123def456-789xyz-00"